requirements.txt
lxml
python-decouple
apache-airflow-providers-apache-spark
requests
pandas
numpy
pyarrow 
fastparquet
scikit-learn
s3fs
